fix: Displays the exercises from a current workout only.
Before we were pulling all exercises from the database but changed code to fit the use case of the coach entering exercises for a specfic workout which doesn't need the database.

This is also in line with the change of recommended exercises being attached to a workout instead of their own collection.